Adding connection to mysql using docker & User profile management

+++ app.py
@@ -5,7 +5,7 @@ from flask_login import LoginManager, login_user, current_user, logout_user, log
 
 app = Flask(__name__)
 app.config['SECRET_KEY'] = "your_secret_key"
-app.config ['SQLALCHEMY_DATABASE_URI'] = 'sqlite:///database.db'
+app.config['SQLALCHEMY_DATABASE_URI'] = "mysql+pymysql://root:admin123@127.0.0.1:3306/flask-crud"
 
 login_manager = LoginManager()
 login_manager.init_app(app)
@@ -53,7 +53,7 @@ def create_user():
     password = data.get("password")
 
     if username and password:
-        user = User(username=username, password=password)
+        user = User(username=username, password=password, role='user')
         db.session.add(user)
         db.session.commit()
         return jsonify({"message": "Usuário cadastrado com sucesso"})
@@ -76,6 +76,9 @@ def update_password(id_user):
      data = request.json
      user = User.query.get(id_user)
 
+     if id_user != current_user.id and current_user.role == "user":
+          return jsonify({"message": "Operação não permitida. Somente administradores podem alterar a senha de outros usuários"}), 403
+
      if user and data.get("password"):
           user.password = data.get('password')
           db.session.commit()
@@ -89,10 +92,13 @@ def update_password(id_user):
 def delete_user(id_user):
      user = User.query.get(id_user)
 
+     if current_user.role != 'admin':
+          return jsonify ({"message": "Operação não permitida. Somente administradores podem deletar usuários"}), 403
+
      if id_user == current_user.id:
           return jsonify({"message": "Deleção não permitida"}), 403
 
-     if user: 
+     if user:
           db.session.delete(user)
           db.session.commit()
 
@@ -102,5 +108,5 @@ def delete_user(id_user):
 
 if __name__ == '__main__':
     app.run(debug=True)
-
+
